# Adds copyStream buffer pool