Reduces memory allocation by reusing byte slices in `copyStream`.

**Profiling results**

```
bin/skipper -inline-routes='* -> inlineContent("Hello") -> <shunt>' -access-log-disabled -enable-profile
```

Baseline (`master`)
```
$ echo "GET http://localhost:9090/test" | vegeta attack -duration=10m -rate=0 --max-workers=500 | vegeta report
Requests      [total, rate, throughput]         16754506, 27924.03, 27923.66
Duration      [total, attack, wait]             10m0s, 10m0s, 8.097ms
Latencies     [min, mean, 50, 90, 95, 99, max]  89.763µs, 14.278ms, 9.62ms, 36.179ms, 43.932ms, 65.012ms, 228.741ms
Bytes In      [total, mean]                     83772530, 5.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:16754506
Error Set:
```
```
$ go tool pprof -top http://localhost:9911/debug/pprof/allocs
...
Showing nodes accounting for 194.44GB, 96.27% of 201.96GB total
Dropped 137 nodes (cum <= 1.01GB)
      flat  flat%   sum%        cum   cum%
  127.97GB 63.36% 63.36%      128GB 63.38%  github.com/zalando/skipper/proxy.copyStream
    9.04GB  4.48% 67.84%    10.13GB  5.01%  github.com/zalando/skipper/filters/builtin.(*inlineContent).Request
    8.16GB  4.04% 71.88%     8.43GB  4.17%  net/textproto.(*Reader).ReadMIMEHeader
       7GB  3.47% 75.35%        7GB  3.47%  net/http.Header.Clone
    5.95GB  2.95% 78.30%    23.89GB 11.83%  net/http.(*conn).readRequest
    5.46GB  2.70% 81.00%     5.46GB  2.70%  github.com/zalando/skipper/proxy.copyHeader
       5GB  2.47% 83.48%     5.12GB  2.53%  github.com/zalando/skipper/proxy.newContext
    4.02GB  1.99% 85.47%     6.21GB  3.08%  net/http.(*Request).WithContext (inline)
    4.01GB  1.99% 87.45%    15.52GB  7.69%  net/http.readRequest
    3.96GB  1.96% 89.41%     3.96GB  1.96%  github.com/zalando/skipper/proxy.(*proxyTracing).logEvent (inline)
    2.48GB  1.23% 90.64%     2.54GB  1.26%  fmt.Sprintf
    2.27GB  1.12% 91.76%     2.27GB  1.12%  net/url.parse
    2.19GB  1.09% 92.85%     2.19GB  1.09%  net/http.cloneURL (inline)
    1.46GB  0.72% 93.57%     1.46GB  0.72%  github.com/zalando/skipper/tracing.CreateSpan
    1.27GB  0.63% 94.20%     1.27GB  0.63%  context.WithCancel
    1.26GB  0.63% 94.82%     1.26GB  0.63%  net.(*conn).Read
    1.19GB  0.59% 95.41%     1.38GB  0.68%  github.com/zalando/skipper/proxy.(*Proxy).setCommonSpanInfo
    1.14GB  0.56% 95.98%     1.14GB  0.56%  bufio.NewWriterSize (inline)
    0.47GB  0.23% 96.21%   176.60GB 87.44%  github.com/zalando/skipper/proxy.(*Proxy).ServeHTTP
    0.13GB 0.064% 96.27%    14.74GB  7.30%  github.com/zalando/skipper/proxy.(*Proxy).applyFiltersToRequest
...
```

This change:
```
$ echo "GET http://localhost:9090/test" | vegeta attack -duration=10m -rate=0 --max-workers=500 | vegeta report
Requests      [total, rate, throughput]         18090343, 30149.97, 30149.93
Duration      [total, attack, wait]             10m0s, 10m0s, 867.153µs
Latencies     [min, mean, 50, 90, 95, 99, max]  74.27µs, 10.584ms, 8.151ms, 23.656ms, 29.943ms, 42.609ms, 132.707ms
Bytes In      [total, mean]                     90451715, 5.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:18090343
Error Set:
```
```
$ go tool pprof -top http://localhost:9911/debug/pprof/allocs
...
Showing nodes accounting for 76.82GB, 97.85% of 78.51GB total
Dropped 100 nodes (cum <= 0.39GB)
      flat  flat%   sum%        cum   cum%
    9.74GB 12.41% 12.41%    10.85GB 13.82%  github.com/zalando/skipper/filters/builtin.(*inlineContent).Request
    8.84GB 11.26% 23.67%     9.10GB 11.59%  net/textproto.(*Reader).ReadMIMEHeader
    7.47GB  9.52% 33.18%     7.47GB  9.52%  net/http.Header.Clone
    6.45GB  8.22% 41.40%    24.42GB 31.10%  net/http.(*conn).readRequest
    5.94GB  7.56% 48.96%     5.94GB  7.56%  github.com/zalando/skipper/proxy.copyHeader
    5.35GB  6.82% 55.78%     5.49GB  6.99%  github.com/zalando/skipper/proxy.newContext
    4.32GB  5.50% 61.28%     6.76GB  8.61%  net/http.(*Request).WithContext (inline)
    4.31GB  5.49% 66.77%     4.31GB  5.49%  github.com/zalando/skipper/proxy.(*proxyTracing).logEvent (inline)
    4.29GB  5.46% 72.23%    16.53GB 21.06%  net/http.readRequest
    2.68GB  3.41% 75.64%     2.70GB  3.44%  fmt.Sprintf
    2.46GB  3.14% 78.78%     2.46GB  3.14%  net/url.parse
    2.45GB  3.11% 81.89%     2.45GB  3.11%  net/http.cloneURL (inline)
    1.63GB  2.07% 83.96%     1.63GB  2.07%  github.com/zalando/skipper/tracing.CreateSpan
    1.36GB  1.73% 85.69%     1.36GB  1.73%  net.(*conn).Read
    1.33GB  1.70% 87.39%     1.34GB  1.70%  context.WithCancel
    1.32GB  1.69% 89.08%     1.52GB  1.94%  github.com/zalando/skipper/proxy.(*Proxy).setCommonSpanInfo
    1.09GB  1.39% 90.47%     1.09GB  1.39%  github.com/zalando/skipper/proxy.(*context).setMetricsPrefix (inline)
    0.86GB  1.09% 91.56%     0.86GB  1.09%  bytes.NewBufferString (inline)
    0.83GB  1.06% 92.61%     0.83GB  1.06%  github.com/zalando/skipper/proxy.appendParams
    0.81GB  1.03% 93.65%     0.81GB  1.03%  github.com/zalando/skipper/routing.(*matcher).match
    0.76GB  0.97% 94.62%     0.76GB  0.97%  context.WithValue
    0.55GB   0.7% 95.32%     0.55GB   0.7%  net/textproto.(*Reader).ReadLine (inline)
    0.54GB  0.69% 96.01%    52.58GB 66.97%  github.com/zalando/skipper/proxy.(*Proxy).ServeHTTP
    0.52GB  0.66% 96.67%     0.52GB  0.66%  github.com/zalando/skipper/logging.NewLoggingWriter (inline)
    0.27GB  0.35% 97.02%     0.80GB  1.01%  github.com/zalando/skipper/metrics.(*CodaHale).MeasureFilterResponse
    0.26GB  0.33% 97.35%     0.81GB  1.03%  github.com/zalando/skipper/metrics.(*CodaHale).MeasureResponse
    0.25GB  0.32% 97.67%     0.78GB     1%  github.com/zalando/skipper/metrics.(*CodaHale).MeasureFilterRequest
    0.14GB  0.18% 97.85%    15.82GB 20.15%  github.com/zalando/skipper/proxy.(*Proxy).applyFiltersToRequest
...
```